+++ packages/compiler/bundler/index.ts
@@ -1,0 +1,3 @@
+import { buildCompiler } from "./build-main";
+
+buildCompiler();